Add files via upload
Insights from my dashboard:
→ There is very small difference in ADR (Average daily rate)generated on
weekdays and weekends, which indicates that hotels are not looking over their
pricing strategy. It should be seen upon to generate more revenue, as the price
factor in this industry is very elastic and changes heavily in weekends and
weekdays.
Direct offline booking platform has the highest overall realization% and ADR, which needs to be worked upon.
More emphasis could be put on booking platforms. ✦ Hotels with lower ratings should look upon the factors behind the low ratings.
There is a strong correlation between occupancy, ratings and cancellation, because of which hotels should check the ratings, especially if they are operating on online channels
Luxury hotels have generated more revenue as compared to business hotels.
Elite room type have generated highest revenue
The highest revenue and rating have been generated by the same hotel in Mumbai, which is also the best performing city among the other cities
Overall, an effective pricing strategy is needed to increase revenue.